🚚 Chore: apiPayload 패키지 global.apiPayload로 이동 #95
