--- src/main/java/callprotector/spring/apiPayload/exception/handler/InvalidCategoryFilterException.java
@@ -1,10 +1,0 @@
-package callprotector.spring.apiPayload.exception.handler;
-
-import callprotector.spring.apiPayload.code.status.ErrorStatus;
-import callprotector.spring.apiPayload.exception.GeneralException;
-
-public class InvalidCategoryFilterException extends GeneralException {
-    public InvalidCategoryFilterException() {
-        super(ErrorStatus.INVALID_CATEGORY_FILTER);
-    }
-}
