+++ src/main/java/callprotector/spring/global/apiPayload/code/BaseCode.java
@@ -1,4 +1,4 @@
-package callprotector.spring.apiPayload.code;
+package callprotector.spring.global.apiPayload.code;
 
 public interface BaseCode {
