+++ src/main/java/callprotector/spring/global/apiPayload/exception/GeneralException.java
@@ -1,7 +1,7 @@
-package callprotector.spring.apiPayload.exception;
+package callprotector.spring.global.apiPayload.exception;
 
-import callprotector.spring.apiPayload.code.BaseErrorCode;
-import callprotector.spring.apiPayload.code.ErrorReasonDTO;
+import callprotector.spring.global.apiPayload.code.BaseErrorCode;
+import callprotector.spring.global.apiPayload.code.ErrorReasonDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
